WinClient-16 Rename TopShelfService to WinService and remove old project
