--- SysHv/SysHv.Client.TopShelfService/NLog.config
@@ -1,16 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<nlog xmlns="http://www.nlog-project.org/schemas/NLog.xsd"
-      xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-      xsi:schemaLocation="http://www.nlog-project.org/schemas/NLog.xsd NLog.xsd"
-      autoReload="true"
-      throwExceptions="true"
-      internalLogLevel="Off" internalLogFile="c:\temp\nlog-internal.log">
-  <targets>
-    <target name="logfile" xsi:type="File" fileName="C:\SysHv.log" />
-  </targets>
-
-  <rules>
-    <logger name="*" minlevel="Info" writeTo="logfile" />
-    <logger name="*" minlevel="Debug" writeTo="logfile" />
-  </rules>
-</nlog>
