--- SysHv/SysHv.Client.TopShelfService/packages.config
@@ -1,7 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="NLog" version="4.6.2" targetFramework="net472" />
-  <package id="NLog.Config" version="4.6.2" targetFramework="net472" />
-  <package id="NLog.Schema" version="4.6.2" targetFramework="net472" />
-  <package id="Topshelf" version="4.2.0" targetFramework="net472" />
-</packages>
