+++ SysHv/SysHv.Client.WinService/NLog.config
@@ -3,7 +3,7 @@
       xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
       xsi:schemaLocation="http://www.nlog-project.org/schemas/NLog.xsd NLog.xsd"
       autoReload="true"
-      throwExceptions="false"
+      throwExceptions="true"
       internalLogLevel="Off" internalLogFile="c:\temp\nlog-internal.log">
   <targets>
     <target name="logfile" xsi:type="File" fileName="C:\SysHv.log" />
