+++ SysHv/SysHv.Client.WinService/Program.cs
@@ -1,20 +1,40 @@
-﻿using System.ServiceProcess;
+﻿using SysHv.Client.WinService.Services;
+using System;
+using Topshelf;
 
 namespace SysHv.Client.WinService
 {
-    internal static class Program
+    class Program
     {
-        /// <summary>
-        /// The main entry point for the application.
-        /// </summary>
-        private static void Main()
+        static void Main(string[] args)
         {
-            ServiceBase[] ServicesToRun;
-            ServicesToRun = new ServiceBase[]
+            //RuntimeInfoGatherer gatherer = new RuntimeInfoGatherer();
+            //Console.WriteLine(gatherer.Gather());
+            /*
+            var systemInfo = new HardwarenfoGatherer();
+            Console.WriteLine(systemInfo.Gather());
+            Console.ReadKey();
+            return;*/
+
+            // here is service setup
+            var exitCode = HostFactory.Run(x =>
             {
-                new SysHvMonitoringService()
-            };
-            ServiceBase.Run(ServicesToRun);
+                x.Service<MonitoringService>(s =>
+                {
+                    s.ConstructUsing(monitor => new MonitoringService());
+                    s.WhenStarted(monitor => monitor.Start());
+                    s.WhenStopped(monitor => monitor.Stop());
+                });
+
+                x.RunAsLocalSystem();
+
+                x.SetServiceName("SysHvMonitor");
+                x.SetDisplayName("SysHv Monitoring Service (Topshelf)");
+                x.SetDescription("Sends telemetry from your computer to system administrator so he can figure out what is going on with your machine when something fails");
+            });
+
+            int exitCodeValue = (int)Convert.ChangeType(exitCode, exitCode.GetTypeCode());
+            Environment.ExitCode = exitCodeValue;
         }
     }
 }
