--- SysHv/SysHv.Client.WinService/ProjectInstaller.cs
@@ -1,24 +1,0 @@
-﻿using System.ComponentModel;
-using System.Configuration.Install;
-
-namespace SysHv.Client.WinService
-{
-    [RunInstaller(true)]
-    public partial class ProjectInstaller : System.Configuration.Install.Installer
-    {
-        public ProjectInstaller()
-        {
-            InitializeComponent();
-        }
-
-        private void serviceInstaller1_AfterInstall(object sender, InstallEventArgs e)
-        {
-
-        }
-
-        private void serviceProcessInstaller1_AfterInstall(object sender, InstallEventArgs e)
-        {
-
-        }
-    }
-}
