+++ SysHv/SysHv.sln
@@ -3,11 +3,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.28729.10
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SysHv.Client.WinService", "SysHv.Client.WinService\SysHv.Client.WinService.csproj", "{EFF890AB-FECB-498C-8555-B701C59A4AAE}"
-EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "client", "client", "{C2705447-0039-4E82-A99C-9D4B6F854C6B}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SysHv.Client.TopShelfService", "SysHv.Client.TopShelfService\SysHv.Client.TopShelfService.csproj", "{982E7E19-5E22-4948-A311-2D5FCAF02EFB}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SysHv.Client.WinService", "SysHv.Client.WinService\SysHv.Client.WinService.csproj", "{982E7E19-5E22-4948-A311-2D5FCAF02EFB}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SysHv.Client.Common", "SysHv.Client.Common\SysHv.Client.Common.csproj", "{F0699901-1F2A-4395-B568-B5725FB1C135}"
 EndProject
@@ -19,14 +17,6 @@ Global
 		Release|x64 = Release|x64
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{EFF890AB-FECB-498C-8555-B701C59A4AAE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{EFF890AB-FECB-498C-8555-B701C59A4AAE}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{EFF890AB-FECB-498C-8555-B701C59A4AAE}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{EFF890AB-FECB-498C-8555-B701C59A4AAE}.Debug|x64.Build.0 = Debug|Any CPU
-		{EFF890AB-FECB-498C-8555-B701C59A4AAE}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{EFF890AB-FECB-498C-8555-B701C59A4AAE}.Release|Any CPU.Build.0 = Release|Any CPU
-		{EFF890AB-FECB-498C-8555-B701C59A4AAE}.Release|x64.ActiveCfg = Release|Any CPU
-		{EFF890AB-FECB-498C-8555-B701C59A4AAE}.Release|x64.Build.0 = Release|Any CPU
 		{982E7E19-5E22-4948-A311-2D5FCAF02EFB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{982E7E19-5E22-4948-A311-2D5FCAF02EFB}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{982E7E19-5E22-4948-A311-2D5FCAF02EFB}.Debug|x64.ActiveCfg = Debug|Any CPU
@@ -48,7 +38,6 @@ Global
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
-		{EFF890AB-FECB-498C-8555-B701C59A4AAE} = {C2705447-0039-4E82-A99C-9D4B6F854C6B}
 		{982E7E19-5E22-4948-A311-2D5FCAF02EFB} = {C2705447-0039-4E82-A99C-9D4B6F854C6B}
 		{F0699901-1F2A-4395-B568-B5725FB1C135} = {C2705447-0039-4E82-A99C-9D4B6F854C6B}
 	EndGlobalSection
